fix: deprecate LISTEN_ADDR in pg_template.yaml
hostname -i` might print multiple IP addresses separated with space,
so it is a bad fit for the listen address. We should better use `*:port` for listen address,
which is a static value. An environment variable with a static value does not help much.
We'd better deprecate and remove LISTEN_ADDR variable.